Renamed FieldCheckbox to FieldCheckboxMultiple
