+++ docs/src/Form/index.js
@@ -32,7 +32,7 @@ class FormExample extends React.Component {
         validationErrorText: 'Password needs to be greater than 8 characters'
       },
       {
-        fieldType: 'checkbox',
+        fieldType: 'checkboxMultiple',
         value: [
           {
             name: 'isManager',
@@ -185,7 +185,7 @@ class FormExample extends React.Component {
         validationErrorText: 'Password needs to be greater than 8 characters'
       },
       {
-        fieldType: 'checkbox',
+        fieldType: 'checkboxMultiple',
         value: [
           {
             name: 'isManager',
